pyproject.toml
[project]
name = "opensoft-25-backend"
version = "0.1.0"
description = "OpenSoft Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.11",
    "langsmith==0.3.11",
    "motor==3.7.0",
    "neo4j==5.28.1",
    "openai==1.66.3",
    "pandas==2.2.3",
    "passlib==1.7.4",
    "pydantic[email]==2.10.3",
    "pydantic-core==2.27.1",
    "pydantic-settings==2.7.0",
    "pymongo==4.11.2",
    "python-dotenv==1.0.0",
    "python-jose==3.4.0",
    "pytz==2025.1",
    "rich==13.9.4",
    "uvicorn==0.34.0",
    "faker==37.1.0",
    "gradio==5.23.1",
]

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "B", "C4", "N", "UP", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.isort]
known-third-party = [
    "fastapi",
    "langsmith",
    "motor",
    "neo4j",
    "openai",
    "pandas",
    "passlib",
    "pydantic",
    "pymongo",
    "python-jose",
    "rich",
    "uvicorn",
    "faker",
    "gradio",
]

[project.optional-dependencies]
dev = ["pre-commit==3.5.0", "ruff>=0.9.3"]
